+++ frontend-dart/lib/models/api_models.dart
@@ -613,6 +613,24 @@ class FriendsResponse {
       );
 }
 
+class Friendship {
+  final String id;
+  final int fromUser;
+  final int toUser;
+  final String status;
+  final DateTime createdAt;
+
+  const Friendship({required this.id, required this.fromUser, required this.toUser, required this.status, required this.createdAt});
+
+  factory Friendship.fromJson(Map<String, dynamic> json) => Friendship(
+    id: json['id'].toString(),
+    fromUser: json['from_user'] as int,
+    toUser: json['to_user'] as int,
+    status: json['status'] as String,
+    createdAt: DateTime.parse(json['created_at'] as String),
+  );
+}
+
 class PendingRequestsResponse {
   final List<Map<String, dynamic>> requests;
