+++ frontend-dart/lib/providers/dynamic_theme_provider.dart
@@ -1,4 +1,4 @@
-import '../core/app_logger.dart';
+import '../core/logging_navigation_observer.dart';
 import 'package:flutter/material.dart';
 import 'package:palette_generator/palette_generator.dart';
 import 'package:cached_network_image/cached_network_image.dart';
