+++ frontend-dart/lib/providers/music_provider.dart
@@ -1,6 +1,6 @@
 import '../core/base_provider.dart';
 import '../core/service_locator.dart';
-import '../core/app_logger.dart';
+import '../core/logging_navigation_observer.dart';
 import '../services/music_service.dart';
 import '../services/track_cache_service.dart';
 import '../services/api_service.dart';
@@ -8,7 +8,6 @@ import '../services/auth_service.dart';
 import '../models/music_models.dart';
 import '../models/result_models.dart';
 import '../models/sort_models.dart';
-import '../services/track_sorting_service.dart';
 
 class MusicProvider extends BaseProvider {
   final MusicService _musicService = getIt<MusicService>();
