+++ frontend-dart/lib/screens/auth/auth_screen.dart
@@ -4,7 +4,6 @@ import 'package:provider/provider.dart';
 import '../../providers/auth_provider.dart';
 import '../../core/theme_utils.dart';
 import '../../core/responsive_utils.dart';
-import '../../core/validators.dart';
 import '../../core/constants.dart';
 import '../../core/user_action_logging_mixin.dart';
 import '../../widgets/app_widgets.dart';
