+++ frontend-dart/lib/screens/home/home_screen.dart
@@ -1,4 +1,4 @@
-import '../../core/app_logger.dart';
+import '../../core/logging_navigation_observer.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../providers/auth_provider.dart';
