+++ frontend-dart/lib/screens/music/track_search_screen.dart
@@ -3,11 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../providers/music_provider.dart';
 import '../../services/music_player_service.dart';
-import '../../services/track_sorting_service.dart';
+import '../../services/music_service.dart';
 import '../../core/theme_utils.dart';
 import '../../core/user_action_logging_mixin.dart';
 import '../../widgets/app_widgets.dart';
-import '../../widgets/sort_button.dart';
+import '../../widgets/track_sort_bottom_sheet.dart';
 import '../../models/music_models.dart';
 import '../../models/sort_models.dart';
 import '../base_screen.dart';
