+++ frontend-dart/lib/screens/playlists/playlist_editor_screen.dart
@@ -8,9 +8,8 @@ import '../../services/websocket_service.dart';
 import '../../models/music_models.dart';
 import '../../models/api_models.dart';
 import '../../core/theme_utils.dart';
-import '../../core/validators.dart';
 import '../../core/constants.dart';
-import '../../core/app_logger.dart';
+import '../../core/logging_navigation_observer.dart';
 import '../../widgets/app_widgets.dart';
 import '../../widgets/scrollbar.dart';
 import '../base_screen.dart';
