+++ frontend-dart/lib/screens/playlists/playlist_licensing_screen.dart
@@ -1,4 +1,4 @@
-import '../../core/app_logger.dart';
+import '../../core/logging_navigation_observer.dart';
 import 'package:flutter/material.dart';
 import '../../providers/friend_provider.dart';
 import '../../core/theme_utils.dart';
