+++ frontend-dart/lib/screens/profile/profile_screen.dart
@@ -9,14 +9,12 @@ import 'package:provider/provider.dart';
 import '../../providers/profile_provider.dart';
 import '../../providers/auth_provider.dart';
 import '../../core/theme_utils.dart';
-import '../../core/validators.dart';
 import '../../core/constants.dart';
 import '../../core/user_action_logging_mixin.dart';
 import '../../widgets/app_widgets.dart';
 import '../base_screen.dart';
 import 'user_password_change_screen.dart';
 import 'social_network_link_screen.dart';
-import 'widgets/music_preference_dialog.dart';
 
 class ProfileScreen extends StatefulWidget {
   final bool isEmbedded;
@@ -892,6 +890,80 @@ class _ProfileScreenState extends BaseScreen<ProfileScreen> with UserActionLoggi
     if (confirmed) {
       auth.logout();
     }
+  }
+}
+
+class MusicPreferenceDialog extends StatefulWidget {
+  final List<Map<String, dynamic>> availablePreferences;
+  final List<int> selectedIds;
+
+  const MusicPreferenceDialog({
+    super.key,
+    required this.availablePreferences, 
+    required this.selectedIds
+  });
+
+  @override
+  State<MusicPreferenceDialog> createState() => _MusicPreferenceDialogState();
+}
+
+class _MusicPreferenceDialogState extends State<MusicPreferenceDialog> {
+  late List<int> _selectedIds;
+
+  @override
+  void initState() {
+    super.initState();
+    _selectedIds = List.from(widget.selectedIds);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: AppTheme.surface,
+      title: const Text('Select Music Preferences', style: TextStyle(color: Colors.white)),
+      content: SizedBox(
+        width: double.maxFinite,
+        child: ListView.builder(
+          shrinkWrap: true,
+          itemCount: widget.availablePreferences.length,
+          itemBuilder: (context, index) {
+            final preference = widget.availablePreferences[index];
+            final id = preference['id'] as int;
+            final name = preference['name'] as String;
+            final isSelected = _selectedIds.contains(id);
+
+            return CheckboxListTile(
+              value: isSelected,
+              onChanged: (value) {
+                setState(() {
+                  if (value == true) {
+                    _selectedIds.add(id);
+                  } else {
+                    _selectedIds.remove(id);
+                  }
+                });
+              },
+              title: Text(name, style: const TextStyle(color: Colors.white)),
+              activeColor: AppTheme.primary,
+            );
+          },
+        ),
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.pop(context), 
+          child: const Text('Cancel', style: TextStyle(color: Colors.grey))
+        ),
+        ElevatedButton(
+          onPressed: () => Navigator.pop(context, _selectedIds),
+          style: ElevatedButton.styleFrom(
+            backgroundColor: AppTheme.primary, 
+            foregroundColor: Colors.black
+          ),
+          child: const Text('Save'),
+        ),
+      ],
+    );
   }
 }
