+++ frontend-dart/test/screens/auth/forgot_password_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:music_room/screens/auth/forgot_password_screen.dart';
-import 'package:music_room/core/validators.dart';
+import 'package:music_room/core/constants.dart';
 
 void main() {
   group('Forgot Password Screen Tests', () {
