+++ frontend-dart/test/screens/auth/signup_with_otp_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
 import 'package:music_room/screens/auth/signup_with_otp_screen.dart';
-import 'package:music_room/core/validators.dart';
+import 'package:music_room/core/constants.dart';
 
 void main() {
   group('Signup With OTP Screen Tests', () {
