+++ frontend-dart/test/services/storage_service_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:music_room/services/storage_service.dart';
+import 'package:music_room/services/auth_service.dart';
 
 void main() {
   group('Storage Service Tests', () {
