+++ frontend-dart/test/services/track_sorting_service_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:music_room/services/track_sorting_service.dart';
+import 'package:music_room/services/music_service.dart';
 import 'package:music_room/models/music_models.dart';
 import 'package:music_room/models/sort_models.dart';
